scripts/requirements.txt
requests>=2.31.0
pyyaml>=6.0.1
python-slugify>=8.0.1
tenacity>=8.2.3
